Fix issues raised on the Automated Testing PR
Document the AbstractIntegrationTest so that we know it serves as a base class for integration tests as it provides database setup.
Add TestCustomUserDetailsService to handle test specific authorisation.
Remove test property file as it is no longer needed because we removed test authorisation logic from production code.
Remove UserDaoImplTest as it heavily depend on mocks on jdbc template and sql queries, we better test this with integration tests