Add properties for the test suite
Added application-test.properties to allow for a separation of application
and test properties using Spring profiles. Documented and refacted
ApplicationConfiguration to check the Spring profiles - this should be
done automatically by Spring but as we are using a property to define
the location of the other property files (under QAT_HOME), it needs to
be done manually. Also added the option to define QAT_HOME as an
environment variable and tested that any property can be overriden with
an environment variable (which will be used in the CI/CD pipeline).