Finish Flyway implementation
Finished the implementation for Flyway for the application itself. It
is currently turned off (via the properties) but when it is enabled we
can put our database scripts into the folder src/main/resources/db/migrate
with a version prefix e.g. `V1__` and they will be automatically run
in the application's database